ai_news_reporter/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-ollama
motor
pydantic
python-dotenv
httpx
tenacity
ddgs